Refactor MountRotation
